refactor: refactor solvers
