+++ txmatching/solvers/ilp_solver/ilp_solver.py
@@ -2,7 +2,7 @@ import logging
 from dataclasses import dataclass
 from typing import Dict, Iterable, Iterator, List, Tuple
 
-from txmatching.solvers.all_solutions_solver.donor_recipient_pair_idx_only import \
+from txmatching.solvers.donor_recipient_pair_idx_only import \
     DonorRecipientPairIdxOnly
 from txmatching.solvers.ilp_solver.solve_ilp import solve_ilp
 from txmatching.solvers.ilp_solver.txm_configuration_for_ilp import \
