+++ txmatching/solvers/ilp_solver/solve_ilp.py
@@ -107,12 +107,12 @@ def _add_static_constraints(data_and_configuration: DataAndConfigurationForILPSo
     for node in data_and_configuration.required_patients:
         ilp_model.add_constr(mapping.node_to_in_var[node] >= 0.5)
 
-    add_debt_static_constraints(ilp_model, data_and_configuration, mapping)
+    _add_debt_static_constraints(ilp_model, data_and_configuration, mapping)
 
-    add_blood_group_zero_debt_static_constraints(ilp_model, data_and_configuration, mapping)
+    _add_blood_group_zero_debt_static_constraints(ilp_model, data_and_configuration, mapping)
 
 
-def add_debt_static_constraints(ilp_model,
+def _add_debt_static_constraints(ilp_model,
                                 data_and_configuration: DataAndConfigurationForILPSolver,
                                 mapping: VariableMapping):
     countries = set(data_and_configuration.country_codes_dict.values())
@@ -128,7 +128,7 @@ def add_debt_static_constraints(ilp_model,
             - mip.xsum(country_giving + country_receiving) <= data_and_configuration.configuration.max_debt_for_country)
 
 
-def add_blood_group_zero_debt_static_constraints(ilp_model,
+def _add_blood_group_zero_debt_static_constraints(ilp_model,
                                                  data_and_configuration: DataAndConfigurationForILPSolver,
                                                  mapping: VariableMapping):
     countries = set(data_and_configuration.country_codes_dict.values())
